aura-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.20
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.9.2
pydantic-settings==2.5.2